Core Client: Avoid Memory Leak by having only single client.
There was memory leak if user index huge number of file. For eg:
When user was indexing around 50k files then for each file it
was building rest client for each request. Consequence is it was
running out of memory and jvm was not able to create further thread
because it reaches 200k thread.
Ideally it should not create these number of thread.  We do not need
to build client for each request. We can just initialize client first
time and later we can use the same.

This patch will create rest client at the time of initialization of
application and when application close it will close the client.

Signed-off-by: Gaurav Kumar Garg <garg.gaurav52@gmail.com>